refactoring contract
